+++ src/App.css
@@ -1,3 +1,3 @@
 .App {
-  margin: 20px;
+  padding: 20px;
 }
