--- src/components/test.js
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function test() {
-  return (
-    <div>test</div>
-  )
-}
